Fix and improve filter_content()
1. Fix a little problem in `filter_content()`: because the regex
which searches for image notes starts with `(?s)` (this makes `.`
match any character, including newline) and uses the greedy `.*`
between `{{ImageNote|...}}` and `{{ImageNoteEnd|...}}`, the regex
matches from the beginning of the first to the end of the last
image note, including everything in between.  Normally this is OK
because image notes are all in one place, but you never know ...
So better use the non-greedy `.*?` and make the search mode really
obvious by replacing the `(?s)` by the explicit flag `re.DOTALL`.

2. Filter out also the contents of collapse boxes.  If people put
part of a nomination discussion into such boxes, they regard
that part as a digression or excursus and not as a part of the
normal voting.  Sometimes people even use collapse boxes to
withdraw complete alternatives in a nomination [1].  We have to
ignore the complete collapse boxes and their contents in all tasks
of the bot.

3. We better escape any braces in a regex.  Python 3.13 doesn’t
complain about the literal braces, of course, but braces clearly
have a special meaning in a regex, so (i) future versions of
Python’s regex engine may very well dislike these plain braces,
and (ii) it’s just good style to escape the braces.

[1] See e.g. this nomination:
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/
File:SW_Ketti_Valley_Nilgiris_Jul25_A7CR_06125.jpg