Filter contents before counting images, sections
Before counting the images or sections in a nomination subpage,
we should apply `filter_content()` to the nomination text in order
to ignore any out deleted, stricken, or commented passages
as well as the contents of collapse boxes.  If people comment out
a part of a discussion, mark it as deleted or put it into a
collapse box, they regard that part as a digression or excursus
and not as a valid part of the normal voting.  Sometimes people
even use collapse boxes to withdraw complete alternative sections
in a nomination [1].  We have to ignore such disregarded parts
for most tasks of the bot.

[1] See e.g. this nomination:
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/
File:SW_Ketti_Valley_Nilgiris_Jul25_A7CR_06125.jpg