Tidy up Candidate.notifyNominator()
As next step I want to refine the handling of the `com-nom` and
`subpage` parameters for the `{{Assessments}}`, `{{FPpromotion}}`
and Co. templates; there are still some corner cases in which
the tests are not correct and the parameter is missing
or added without being necessary.  But it is rather tricky
to change the related code in `Candidate.notifyNominator()`
because the flow of control in that method is quite unclear;
it duplicates similar code (e.g. for committing the changes)
instead of using the same code in a well-defined manner.
This state would urge us to duplicate any new/changed code, too.

Therefore let’s better tidy up this method first and rewrite it
in a well-structured manner with transparent flow of control.
At this opportunity let’s also improve the names of variables,
the comments and the messages generated by the method to make
every step completely clear and to avoid any misunderstandings.

This also fixes some minor issues with the method.  For example,
the old test if the `{{FPpromotion}}` for the current candidate
is already present on the talk page just failed if the template
used the `subpage` parameter; so the bot could add the template
a second time to the talk page.