Improve handling of moved user talk pages
Sometimes user talk pages are redirect pages -- e.g., if the
username has been changed or if the user wants visitors to leave
notes on another talk page instead.

`Candidate.notifyNominator()` and `notifyUploaderAndCreator()`
read the user talk page with `talk_page.get(get_redirect=True)`.
But if the talk page is actually a redirect, this means we get
the wikitext of the redirect page and add the promotion template
to it -- which is obviously the wrong place, as nobody sees it.

Let’s improve this by reading the talk page with the parameter
`get_redirect=False`.  This way the Pywikibot framework raises
an `IsRedirectPageError` exception; we can handle it and try
to resolve the redirect.  If this is successful, we can add
the message to the correct user talk page.  If it fails,
we print a useful warning to the CLI resp. to the log file.
We use a warning, not an error, and we do not call `ask_for_help()`
because this is ‘just’ the user notification.

In `notifyUploaderAndCreator()`, things are a bit more complicated.
We must handle the image files from a set nomination one by one
because they may have been uploaded by different users.
That’s very unusual and discouraged by the current FPC rules,
but the bot stills supports that special case.  So the uploader
is often the same for all or at least for several images,
but not always.  Therefore let’s save each resolved redirect
in a simple dictionary and check it when handling the next file.
At the first glance this looks complicated, but it is much
faster than resolving the redirect again.