Always add the ‘com-nom’/‘subpage’ parameter
Until now, the `com-nom` parameter of the `{{Assessments}}` template
and the `subpage` parameter of the user notification templates
`{{FPpromotion}}`, `{{FPpromotionSet2}}` `{{FPpromotedUploader}}` ...
was added if and only if the name of the nomination subpage differed
so significantly from the name of the promoted image that the
templates were not able to link automatically to the nomination.
(The templates use some really nifty code for that, see especially
the auxiliary template `{{FPC/}}`, `{{FPCR/}}`, and `{{FSC/}}`.)

One the one hand that was a cool and elegant solution because it kept
the wikitext of the templates on the file description and user talk
pages as simple and short as possible.  On the other hand it made
the code in FPCBot which checks whether we need the parameter or not
unnecessarily complicated, and, more important, there were always
some corner cases in which that check failed, i.e., the parameter
was not added but despite that it would have been necessary.
Therefore e.g. the templates did not link correctly to repeated
nominations with their ‘.../2’, ‘.../3’ etc. suffix.

But the most important problem is that FPs are renamed quite often
and file movers, including experienced ones, almost never remember
that they *must* add the `com-nom` parameter to the `{{Assessments}}`
template after renaming a FP.  I have fixed many broken FPC links,
but honestly I am getting tired.  It does not help much to remind
file movers to add the `com-nom` parameter ...  They forget it again.

In a discussion on the Featured pictures talk page in August 2025,
it was pointed out it is “just a convenience for the human” that
these parameters are optional and are worked out by the templates;
but there is no inconvenience if the bot always adds them.

This means that it would be a waste of time to fiddle with the code
in FPCBot that decides whether we need the parameters or not.
The best solution is a coarse, but solid one: let’s just *always*
add the params, so the templates will not break even if the FPs
are renamed after the promotion.  (The only remaining case which
necessitates an update of the parameters is if somebody renames
the nomination subpage itself.  But that’s quite rare and normally
done only by some maintainers who should know what they are doing.)

This commit also fixes a related problem.  The value for all these
parameters was generated with `self.fileName(alternative=False)`.
But `Candidate.fileName()` is a method with too many duties,
therefore it does not always return the correct value as needed
for the parameters -- esp. it removes the ‘.../2’, ‘.../3’ suffix
because it would be irritating in other use cases, but the
`com-nom` and `subpage` parameters must include that suffix.
To get exactly the parameter values we need and to take the load
of `Candidate.fileName()`, we add a new tailor-made method,
`Candidate.subpageName()`, which delivers the correct values.

The commit also improves the names of some related variables and
the wording of some related comments.  Finally it fixes a bug
in `Candidate.addAssessments()`: we must add a `|` at the start
of the parameters string if it does not already start with a `|`,
but our test for that condition fails if the string begins with
some whitespace or is empty; so call `params.strip()` and then
test if the string is still not empty before checking whether
the first character is already a `|` or not.  (That’s indeed
relevant because people sometimes format the wikitext of the
`{{Assessments}}` template by adding fancy whitespace between
the individual parameters.)