Save pages only if the text has really changed
In `DelistCandidate.removeFromGalleryPages()` as well as in
`DelistCandidate.removeAssessments()`, call `commit()` only
if the modified wikitext of the page really differs from
the original text.  Of course this makes no difference for
the results, but if you run the bot program interactively
in the CLI in order to test the code and to find bugs,
it is quite confusing when the program asks you again and
again whether it should commit some changes, but actually
there are no changes ;–).

Improved the edit summary used in `removeAssessments()` by
adding a link to the delisting nomination subpage.