Straighten fileName(), drop cleanTitle/cleanSetTitle()
Thanks to the last commit `Candidate.fileName()` is no longer a vague
method with different purposes, but has a clear purpose: it retrieves
as good as possible the name of the candidate image of the nomination.
Therefore it must always consider the alternative, if selected,
and always include the file extension, so we can drop the parameters
which controlled them.  On the other hand we can improve some points,
e.g. make sure to remove any ‘/2’, ‘/3’ suffix from repeated noms
*before* checking whether a file with that name exists or not
(until now that test always failed for all repeated nominations).
So this becomes are more specialized method which fulfills its
single purpose better.  (We still have to improve the error and
exception handling, this should be done in a separate commit.)

After expanding the new `Candidate.subpageName()` method with an
additional parameter `keep_number` (which controls whether a trailing
‘/2’, ‘/3’ is kept or removed), we can use this method also to replace
all uses of `Candidate.cleanSetTitle()`: that method is superfluous
now thanks to the `subpageName()` method.

Likewise we can also drop `Candidate.cleanTitle()` because depending
on the purpose every usage can be replaced either by calling
`Candidate.subpageName()` or by getting the filename from `fileName()`
resp. `setFiles()` and applying `bare_filename()` to it.

These changes do not only remove two methods, but also sharpen the
profile of the remaining methods `fileName()` vs. `subpageName()`.

A positive side-effect of this commit is that we do not longer use
the name of the set nomination as caption for all images from a set
when adding them to a gallery page, but use the individual filenames
as the base for the captions; therefore they are more precise now.