Get DelistCandidate.removeFromFeaturedLists() running
The method `DelistCandidate.removeFromFeaturedLists()` did not work
at all because it found neither the FP gallery page which contains
the former FP nor the page from the chronological archive with the
entry for the former FP.  The `Page.getReferences()` method from
the Pywikibot framework is just the wrong point to start with;
it omits some of the pages which use an image, AFAICS all pages
which use the image within a `<gallery>...</gallery>` element,
and both the FP gallery pages and the chronological archive pages
always use `<gallery>...</gallery>` elements.  The correct method
is `FilePage.using_pages()`, the iterator returned by that method
yields also pages which use an image within a `<gallery>`.

Changing the code in this way fixes a long-standing bug which
prevented that FPCBot did its job with ‘delist’ nominations.

We use the parameters `namespaces=["Commons"]` and
`filterredir=False` when calling `FilePage.using_pages()`;
so we get just pages from the ‘Commons:’ namespace (we only
need these pages) and omit any redirect pages, fine.