Fix counting, prevent parking of set delisting/D-&-R
Fixed the vote counting and prevent the accidental parking of
set delisting nominations as well as delist-and-replace nominations.
We must implement proper support for both of them separately.
For now, we just make sure that the bot does not cause any
damage to such nominations:

1) When counting a set delisting or delist-and-replace nomination,
use ‘X’ for all three vote counts and add a hint that the nomination
must be counted and processed by hand.  This is very similar to
what we already do when closing a FP nomination with alternative(s).

2) If a user has counted and reviewed a delisting nomination or
delist-and-replace nomination, we must not accidentally park it
like a simple delisting nomination, this would just leave a lot
of mess behind.  Therefore cancel the parking procedure early
and add a request for help to the FPC talk page.