Rename and improve removeFromFeaturedLists()
Renamed the method `removeFromFeaturedLists()` of the
`DelistCandidate` class to `removeFromGalleryPages()`,
similar to the already renamed method `addToGalleryPage()`
of the `Candidate` class, to make the purpose and realm
of this method explicit.

Improved the docstring and the comments and added some more
structural comments to document the purpose of this method
as well as the meaning of the individual tasks clearly.

Made the nested `for ... if ... if ...` flatter by using
`continue` if the page name does not meet our requirements.
Bundled the reading of the gallery/archive page in one place.
Added exception handlers and appropriate error messages
for reading/writing errors about the gallery/archive page.
Added some more status messages which are printed to the
CLI/Log.  All these changes should help with debugging the
new code as well as with finding page errors during the
normal usage of the bot.

Added a test if the chronological archive entry is already
marked as delisted (this can happen if the parking procedure
has previously been interrupted), to avoid double ‘Delisted’
hints.  Added error messages which are used if marking the
image as delisted/removing it did not work.

Linked the keyword ‘Delisted’ (which is added to the entry
in the chronological archives) to the relevant delisting
nomination.  This allows users to jump from the entry to
the delisting nomination with a single click and makes it
much easier to comprehend the reasons for the delisting.

Added the application of `str.rstrip()` to each entry from
the chronological archives because many entries there end
with a trailing space; the `str.rstrip()` avoids that we get
two spaces before the ‘Delisted’ hint.