Improve removeAssessments()
Dropped the code that handled the old `{{Featured picture}}`
template: there is not a single FP on Wikimedia Commons which
still uses that template, so we can ignore it.

We can improve the updating of the `{{Assessments}}` template
if we first extract its parameters with a constant general regex,
update the parameters and replace them afterwards.
This general regex uses the `re.DOTALL` flag because sometimes
people break the wikitext of the template into several lines,
therefore we must use the flag to find and update such templates.
The regex which is used to change the value of the `featured`
parameter from `1` to `2` is easier to understand if we use
the `\g<1>` notation for a backreference to group 1 and therefore
do not need anymore to use the octal character code for the `2`.
We must replace at most one instance, so use `count=1`.

We must update also the `com-nom`/`subpage` parameter.
The `{{Assessments}}` is clever enough to link specifically
to a `/removal/` nomination if the value of `featured` is `2`,
but this will not work e.g. if the image has been renamed
since the original FP promotion and the name of the delist
nomination therefore differs from the name of the original
FP nomination.  So it is necessary to update the parameter,
especially now that we *always* add the `com-nom` parameter.
We use the opportunity to replace the old `subpage` parameter
by the modern `com-nom` parameter.  Otherwise the replacement
of that parameter works similar to the updating of the
`featured` parameter.  Again we use the `\g<2>` notation
because it is possible that the name of the nomination subpage
begins with a figure, so it would confuse the regex engine
when appearing right after `\2`.

We must also remove the ‘Featured pictures of/from/by ...’
categories, without touching project-specific categories like
‘Featured pictures on Wikipedia, <language>’.  This is tricky,
and there are also some Commons FP categories with irregular
names like ‘Featured maps’, ‘Featured night photography’ (sic),
or ‘Featured black and white photographs’. Therefore there is
no perfect solution and we use a compromise with three queries
to find > 90% of the categories that must be removed, without
too many false positives.