Remove method Candidate.getImagePage()
Removed the simple method `Candidate.getImagePage()`.
It is not used anymore anywhere in the code.  Furthermore,
whenever we want to get the image description page, we also
need the filename of the image for messages, logging, etc.,
and if we already have the filename it’s just a single line
of code to get the page, too.  And if we use `self.fileName()`
to get the filename and then call `self.getImagePage()`,
we are calling `self.fileName()` two times in the end.
Therefore it does not make much sense to have a separate
method for this.