Move FP parking methods to class FPCandidate
Almost all methods with handle the parking procedure for successful
FP candidates are used only by the subclass `FPCandidate`
(well, that makes sense, doesn’t it? ;–), with the sole exception
of `moveToLog()` which is also used by subclass `DelistCandidate`.
But for historical reasons all of these methods are still part of
the base class `Candidate`.  Let’s move these methods (naturally
except `moveToLog()`) to the subclass `FPCandidate`.  This improves
the structure of the program and helps to understand it more easily.
It also makes explicit that these methods cannot be used by the
other subclass, `DelistCandidate`.

Besides moving the definitions of these methods, adding a docstring
and tweaking another one, this commit does not make any changes
to the code, i.e., it is functionally completely equivalent
to the previous state.

Please note that the usual `diff` tools, including the view
of this commit on GitHub, cannot display correctly that this
commit has actually changed nothing but just moved the methods.