Insert new FPs at beginning of gallery sections
As proposed by Wikimedia Commons user Tagooty on 2025-08-03
at the FPC talk page and discussed there, it makes sense
to insert new featured pictures at the beginning of the selected
gallery section.

As explained by user W.carter in the same discussion, originally
the gallery sections were sorted just chronologically, but later
this chronological order has been eliminated during countless
changes and refinements of the gallery structure.  In addition,
many gallery sections are sorted systematically by taxon (in the
‘biology department’ of the galleries), by year (on the ‘historical’
pages), by place name or geography (on many of the ‘places’ pages),
etc.  This kind of sorting must be done by editors, but the bot
can help them by inserting new images at the beginning of each
section to emphasize that they must be sorted.  Visitors profit
from this new placement, too, because they will typically look
at the first images in each section and should find there the
newest images; and returning visitors will appreciate that they
can see easily what is new in each section.  So by this change
we loose nothing and gain several advantages.

Because we do not need anymore a `.+?` which runs over the contents
of the complete `<gallery>...</gallery>`, but only have to search
till the end of the opening `<gallery ...>` tag, we can use `[^>]*`
instead of `.+?` (allowing `<gallery>` tags without any attributes)
and do not need the `re.DOTALL` flag anymore.

Sometimes there is an empty line after the opening `<gallery...>`,
sometimes there is not even a NL between `<gallery...>` and the
first entry.  To achieve a correct and uniform formatting,
made sure that there is exactly a single NL after `<gallery...>`
by skipping any original whitespace at that place and adding
a `\n` instead.

However, we still append ‘unsorted’ FPs to the very end of the
‘Unsorted’ section because that one is a special section with
no particular meaning besides, well, being ‘unsorted’.