Rename instance variable ‘page’ to ‘_page’
All instance variables (properties) of the `Candidate` class and its
subclasses are flagged as private by starting their names with `_`,
with the sole exception of the `page` variable.  But it makes sense
to mark this variable as private, too -- not for security, but for
safety reasons, i.e. to make clear that the value of this variable
should not be changed from code outside of the class.  Therefore
renamed that variable to `_page`.

To provide read-only access to the value of `_page`, added the
usual `@property` getter `page`.  So any existing code which reads
the value of that instance variable works just like before.