Add ‘Commons featured picture’ claim (Q63348049)
Added a new method to the `FPCandidate` class which extends
the parking procedure of successful FP candidates by a new step:
to add the ‘Commons quality assessment’ (P6731) claim
‘Wikimedia Commons featured picture’ (Q63348049) to the
Media Info (structured data) of the new featured picture,
resp. of all files in a successful set nomination.

This is a logical complement for our parking procedure.  Of course
it was not missed because that structured data claim is added
automatically by BotMultichillT to every new Featured picture;
but FPCBot should be able to do this itself, after all that claim
is an integral part of the Featured picture promotion.

NB: There is a little issue with saving of changes to the Media
Info (structured data) in `commit_media_info_changes()`.
Whenever this is called a 2nd time or after another change,
I get a Pywikibot warning:
    ‘WARNING: API error badtoken: Invalid CSRF token.’
This appears to be harmless (the change is still saved),
but we must keep an eye on it.  Maybe there is something we can
optimize, maybe this is related to some Pywikibot bugs, cf. e.g.:
    https://phabricator.wikimedia.org/T261050