Improve gallery link cleaning
1. When the bot counts the votes in a finished nomination, it adds
the preliminary results template, including the gallery link.
Let’s polish the gallery link a little bit by removing leading or
trailing whitespace and replacing underscores or non-breaking spaces
by plain spaces (underscores are frequent because users just copy
the gallery link, NBSP can be entered by accident with some keyboard
settings, e.g. on macOS or Linux).  This makes it easier for the
reviewing user to read and check the gallery link and it makes
the whole results template look prettier and more professional.

2. We must still clean the gallery link a second time when we
analyse it in order to add the new FP to the list of new FPs and to
the appropriate gallery page, because it is possible that users
change the gallery link and add again leading/trailing whitespace,
underscores or non-breaking spaces.

It makes more sense when we move this cleaning one level up,
from the method `FPCandidate.addToGalleryPage()` to the method
`FPCandidate.handlePassedCandidate()`.  This way the result is
used also by `FPCandidate.addToFeaturedList()`, guaranteeing
that we get a clean gallery link there, too.

Of course this never was a problem because `addToFeaturedList()`
uses only the basic part of the gallery link and that part
traditionally doesn’t contain any spaces etc., but is a single word.
But it is quite possible that this is changed at some point,
and then the search for the correct section on the page with
the most recent FPs could fail without cleaning the link before.
So let’s be just on the save side and use the cleaning of the link
which we need to do anyway for `addToGalleryPage()` right in
`handlePassedCandidate()`, then the result can be used everywhere.

3. Split out the cleaning as a new utility function and call
that function from both places.