Improve FPCandidate.addToFeaturedList()
The usage of a single simple `re.sub()` replacement in the old
implementation of `FPCandidate.addToFeaturedList()` was ingenious,
but had some disadvantages which make it advisable to replace it
by a new, more pedestrian, but also more stable implementation.

The old code failed if there was no subheading with the specified
gallery name, but also if there was just the slightest formatting
discrepancy either in the subheading or in the following `<gallery>`
element (e.g., less than or more than exactly four entries).
This failure went unnoticed, because it just had the consequence
that the section was not changed at all.  Therefore some users
(which watched the list of recently promoted FPs) complained again
and again why *their* new FP was not listed, but many other failures
were never noticed.

The new implementation first searches for the matching subheading
plus `<gallery>` element; it adds both an error message to the log
and a request for help to the FPC talk page if it can’t find
that section.  This means any errors will be noticed, and new FPs
can be added manually to the list of recent FPs if necessary.
It also means that tiny formatting discrepancies in the list are
tolerated now and more important discrepancies are reported.

An additional advantage of the new implementation is that it does
not depend on the exact count of entries in the `<gallery>` element;
it just removes all but the newest three entries and adds the new
entry before them, and works even if the gallery element has less
than four entries or is completely empty.  This makes the new code
more flexible and less error-prone, too.

The new implementation adds the bare name of the new FP as a label
(caption) to its entry.  This makes the list of recent FPs easier
to understand (user can use the hover feature to see the labels)
and brings it in line with all other FP gallery pages (all of them
already use captions and show them when hovering an entry).
We also improve the edit summary which is recorded in the history
of the list page by mentioning the name of the section.

On errors reading the list page, the new implementation does not
crash anymore (like the old implementation), but prints an error
to the log and a request for help to the FPC talk page.