Format -close output like -info/-park output
`Candidate.closePage()` is the only method which uses double ASCII
quotation marks for the demarcation of filenames in the output;
the very similar tabular output generated by `printAllInfo()`
and `park()`, i.e. by the `-info` and `-park` tasks, does not use
quotation marks, but a colon after the filename.  This makes no
difference for the usage of the program as bot program, but when
you run it locally in order to test the code or to check some
nominations, the formatting difference is quite irritating.
Therefore changed the format to match the output formatting used
in `printAllInfo()` and `park()`.

Added an exception handler to catch any errors during the reading
of the FP list page and to report them properly both with an error
message for the CLI/log and with a request for help on the FPC
talk page (if reading the FP list page fails, something is really
wrong or the setup of the FP presentation has changed drastically,
in both cases we need the attention of human editors).

Improved the docstring and the structuring of the method with blank
lines and structural comments to make it even easier to understand
what is done here and why it is done in that order.