Improve 5th day handling and edit summary for closing
The Rules of the fifth day do not apply to nominations which
contain alternative images.  This was checked in the method
`Candidate.closePage()`, but not in `Candidate.statusString()`,
so the output of the latter method was inaccurate.  In order to
fix this once and for all and for all possible uses of the method,
moved the test whether the nomination contains alternative images
or not right to `Candidate.rulesOfFifthDay()`, so it is applied
*always* and can be removed from `Candidate.closePage()`.

The hint whether a nomination was closed due to the Rules of the
fifth day or not was appended to the edit summary, but not in the
same parenthesis as all other information, but like an oddly
formatted supplement in another parenthesis.  Integrated it into
the main parenthesis and used uniform formatting.

There is no reason to use the code-like formatting ‘featured=yes’
and ‘FifthDay=yes’ in the edit summary; it’s a summary for human
editors, not pseudo-code for the bot, so it should be easy to
understand for normal human readers.  Therefore changed the format
to ‘featured: yes, 5th day: yes’ which IMHO is both clearer and
easier to read.