Close withdrawn/FPXed noms regardless of alternatives
The bot closes nominations which are marked as withdrawn or FPXed/
FPDed when one full day has passed since the last edit.  But until
now there was an exception: if the nomination contains alternative
images, the status as withdrawn or FPXed was ignored so that
the nomination would run for full nine days (the Rules of the
5th day are ignored for nominations containing alternatives, too).

Is this exception correct?  I don’t think so.  If a nomination is
withdrawn or FPXed, it should be closed (after waiting a day since
the last edit) even if it contains alternative images.  Maybe this
exception was made for the case that a user wants to withdraw just
a single alternative; but this is not how we handle it in practice.
Because it’s not possible to distinguish the withdrawal of a single
alternative clearly from a complete withdrawal, we understand the
`{{Withdrawn}}` template always as a withdrawal of the complete
nomination.  Likewise it’s obvious that the `{{FPX}}` and `{{FPD}}`
templates always concern the complete nomination.

Therefore I remove the code for that exception.