Catch and report failed Wikidata connection
When we want to add the ‘Commons quality assessment’ (P6731) claim
‘Wikimedia Commons featured picture’ (Q63348049) to the Media Info
(structured data) of one or more new FP(s), we must create a new
Pywikibot `page.Claim` object.  This in turn requires a Pywikibot
`Site` object for Wikidata, but sometimes creating it fails due to
connection errors [1].

Until now the exception raised by Pywikibot in this case was not
handled in our code and stopped the complete run of the bot program.
This seems unreasonable because the FP claim is just a minor part
of the parking procedure.  So let’s catch that exception (and any
other exceptions which are raised during the creation of the `Site`
object for Wikidata) , report the error, and skip this step of the
parking procedure, handing it over to manual handling by adding
a specific request for help to the FPC talk page.

[1] As reported e.g. here:
https://commons.wikimedia.org/w/index.php?title=Commons_talk:
Featured_picture_candidates&diff=prev&oldid=1078899882