In help requests, print exceptions more nicely
We often refer to certain exceptions when adding a request for help
to the FPC talk page.  Let’s make these references more useful and
easier to read by (1) adding the type/name of the exception as an
useful hint; (2) citing the exception message in italics; and also
(3) by removing the period from the end of many Pywikibot exception
messages (the period is irritating because it breaks the grammar of
our message).

BTW: Improved the wording of some requests for help.