Use ‘(edit) summary’ instead of ‘message’/‘comment’
The official term of the Mediawiki software for the short comment
which is used when saving the changed contents of a page is
‘edit summary’ – see
    https://www.mediawiki.org/wiki/Help:Edit_summary
Therefore let’s use this term in all places instead of the misc.
words we used before (‘comment’, ‘message’, ‘summary’, ...).
This makes explicit that we are referring to the Mediawiki edit
summary in all these places, and not e.g. to a git commit message
or something else.