Add DelistCandidate.removeFromFeaturedList()
Until now we omitted the first step of the parking procedure
for delisting candidates, namely to remove the delisted FP
from the list with recently featured images that is used on
the FP landing page.  Indeed this is normally not required.
A comment (which is removed in this commit) used to state:
“We skip checking the FP landing page with the newest FPs;
the chance that the image is still there is very small,
and even then that page will soon be updated anyway.”

That’s correct.  But some sections of the list are updated
only very rarely (e.g. the ‘Other lifeforms’ section which
contains FPs promoted two years ago), so a delisted FP
could stay there for years.  That would be bad.  Furthermore,
removing a FP from the list is easy.  So let’s just do it.