Create P580 value right away with all-zero time
When I implemented the generation of the timestamp value for the
FP assessment claim, I wanted to use Pywikibot’s tools because
the resulting timestamp is, well, used with a Pywikibot method.
But after reconsideration I think it’s ridiculous and inelegant
to format the timestamp first and then to use a regex in order
to nullify the time part.  So here we gain nothing from using
the special Pywikibot stuff, we can simply use Python’s native
`datetime.datetime` objects and format the timestamp right away
with an all-zero definition for the time part.  The result
(you guess it) is exactly the same, and the new code is simpler,
shorter, and easier to understand.