Ping nominator on gallery link problems
When I implemented the `ask_for_help()` function, I did not want
to blame any users, so I did not include the name of the nominator
or anything similar in the messages.  But while the feature works
nicely and is received well, we still have the issue that esp.
gallery link problems are often not fixed soon, and normally not
by the nominator (who should be responsible to fix them).

Therefore added a new option to ping the nominator when the bot
adds a request for help to the FPC talk page, and enabled it
(for now) with some common gallery link problems.