Handle {{FPD}} templates like {{FPX}} templates
Until now the bot program did handle the `{{FPD}}` template like
the `{{Withdraw}}` template.  But the `{{FPD}}` template is not
added by the nominator, but by an observant user; it is more or
less a special variant of the `{{FPX}}` template, used for the
special case that a nomination is invalid because the nominator
already has two open nominations.  From the bot’s point of view
the meaning of `{{FPX}}` and `{{FPD}}` templates is equivalent.

Therefore it makes more sense to handle the `{{FPD}}` template
together with the `{{FPX}}` template, i.e. with the same regex and
the same method, not as a variant of the `{{Withdraw}}` template.
Changed the code and amended the comments and status strings to
reflect properly whether a nomination has been withdrawn or was
FPXed/FPDed.

In order to allow users to cancel both templates by commenting them
out or by using `<s>...</s>` etc., amended the method `isFPX()`
by first applying `filter_content()` to the text of the nomination.
Streamlined the method `isWithdrawn()`.

Improved the regexes `FpxR` and `WithdrawnR`; both should tolerate
optional whitespace after the ‘FPX’/‘FPD’, we better escape braces,
and there is no reason to use the `re.MULTILINE` flag because it
influences only the meaning of `^` and `$` which are not used here.
We also need to support `{{Wdn}}` which is an official abbreviation
for `{{Withdraw}}`.

In `Candidate.park()`, it’s not the best solution to apply the check
for withdrawn, FPXed or FPDed nominations only *after* looking for
verified results.  First, this is not very logical (if a nomination
was withdrawn etc., counting votes and adding results does not make
any sense anymore); second, it had the consequence that withdrawn,
FPXed or FPDed candidates were labelled with ‘no verified results’
in the CLI/log output from that method; quite confusing, because one
would expect them to be labelled more exactly as ‘withdrawn’ etc.
Therefore let’s check first whether a candidate has been withdrawn
etc. or not, and search for a verified result only if the candidate
was neither withdrawn nor FPXed or FPDed.

Consider `{{FPX}}`/`{{FPD}}` also when printing the overview of all
open nominations with `printAllInfo()` (for now subsuming it under
the ‘W’ column, just like `{{Withdraw}}`); and also when generating
the status info with `statusString()`, giving a clear hint why a
nomination is marked in the ‘W’ column of `printAllInfo()` output.
Until now the output of the `-info` task did list such nominations
as ‘Active’, ‘Featured’ or ‘Not featured’, that was quite confusing.