Polish and document class ThreadCheckCandidate
Added docstrings to the class `ThreadCheckCandidate` and to its
methods.  Used the `super()` construct instead of referring
explicitly to the name of the superclass, improving flexibility
and legibility of our code.  Prefixed the names of both instance
variables with `_` to make clear that they should not be used
from outside of the class.