Document custom SIGINT signal handler
Added a docstring to `signal_handler()`, renamed the parameter
`signal` to `signal_number` (because we actually get a number,
not some special kind of signal object, and in order to avoid
redefining/shadowing the name `signal` from the outer scope),
and added a comment to the installation of our signal handler.
Added a structural comment to the common entry point idiom.

NB: Of course this is just a kind of polishing.  Actually I am
not sure whether this is the best way to handle SIGINT/Ctrl-C;
I have done some testing and comparisons with Python’s default
handling of SIGINT/Ctrl-C, but the results are inconclusive.
I will look into this more thoroughly later.  For now I just
wanted to fix the redefining/shadowing of the name `signal`
from the outer scope and to add some basic documentation.