Tidy up precompiled regex constants
Improved the order of the constants a little bit to group them
more clearly according to their use, and added/removed empty lines
in order to make groups of related constants more obvious.
Split some overlong strings into several lines in order to make
it easier to read and edit them.  Removed the pointless extreme
indentation from the definition of `VerifiedResultR`.
Escaped all braces in regex strings properly, as recommended.
Improved the comments which explain the use of the constants.

Some of the precompiled regex constants use the `re.MULTILINE` flag
without reason: that flag changes only the meaning of `^` and `$`,
if a regex does not use these special characters the flag is useless.
So removed it from these constants in order to simplify the code.

Changed `.*` at the end of `VerifiedResultR` to `.*?` to make sure
that we stop at the first instance of `}}` -- previously the regex
captured everything until the last instance of `}}` on the same line
which can give incorrect results if a user puts other template(s)
on the same line (sometimes users do weird things ;–).  Amended
the negated character classes used when searching for the values
of `gallery` and `alternative` by excluding the NL character, too.

Formatted and commented `VerifiedDelistResultR` in the same way
as `VerifiedResultR` in order to make it easier to understand and
to document the meaning of all regex subgroups.

The compiled regexes `CountedTemplateR`, `DelistCountedTemplateR`,
`ReviewedTemplateR`, and `DelistReviewedTemplateR` are never used
-- as the comment claims -- to ‘strip away’ ‘the entire line’
(and there is also no reason to do this).  So we can remove both
the misleading comment ‘Matches the entire line including newline
so they can be stripped away’, the stuff at the beginning and end
of these regexes, and the `re.MULTILINE` flag.  This makes these
regexes much simpler and therefore faster, less prone to errors,
and easier to understand.

Dropped the precompiled regex constant `LastImageR`, it is not
used anywhere in the code and I also do not see an obvious use
case for it.